refactor(cv): reorganize cv module structure and update imports
