+++ braincell/_cv/base_test.py
@@ -18,9 +18,9 @@ import unittest
 import brainunit as u
 import numpy as np
 
-from braincell.cv import CV, CVPerBranch
-from braincell.cv._cv import build_cvs
-from braincell.cv._lower import default_paint_rules
+from braincell._cv import CV, CVPerBranch
+from braincell._cv.base import build_cvs
+from braincell._cv.lower import default_paint_rules
 from braincell.filter import RegionMask
 from braincell.morph.branch import Branch
 from braincell.morph.morphology import Morphology
