+++ braincell/_multi_compartment/clamp_table.py
@@ -6,14 +6,12 @@ the legacy ``Cell._point_clamp_input`` used, and lets
 """
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING
 
 import brainunit as u
 import numpy as np
 
-if TYPE_CHECKING:
-    from braincell.compute._point_tree import PointTree
-    from braincell.compute._runtime import MechanismLayout
+from braincell.compute._point_tree import PointTree
+from braincell.compute._runtime import MechanismLayout
 
 __all__ = ["CLAMP_KINDS", "ClampActiveTable", "build_clamp_active_table"]
